Revert "Reland "Changing PaintFlags get/setAlpha to use only floats""
This reverts commit d3bf6bf2426cf9e108646dfb3632e04e2750e3ce.

Reason for revert: css3/blending/svg-blend-multiply-alpha.html is failing now after this CL:
https://ci.chromium.org/p/chromium/builders/ci/mac12-arm64-rel-tests/647
https://ci.chromium.org/ui/p/chromium/builders/ci/linux-chromeos-rel/57889

Original change's description:
> Reland "Changing PaintFlags get/setAlpha to use only floats"
>
> This is a reland of commit 3cd7b9e34437a6babeeae8ed2184cc002ac32941
>
> There was an issue with the ash build, particularly with the file
> ash/assistant/ui/logo_view/logo_view_impl.cc that had some missing
> PaintFlags::setAlpha call.
>
> I'm also addressing other comments of changing 255 for 255.0f to
> avoid conversions.
>
> I had to rebase two tests that look exactly the same but there is
> probably a small discrepancy in the RGBA values, probably the alpha has slightly changed with a small enough difference to not be
> noticeable.
>
> Original change's description:
> > Changing PaintFlags get/setAlpha to use only floats
> >
> > This CL is part of the effort of increasing color precision and
> > color gamut support in chrome.
> > https://docs.google.com/document/d/1y5qhBuK7sFkXt7c6vcrAJeIofj3h2UQ3btX3Hv1hrz8/edit#heading=h.69111d60ji6h
> > particularly
> > https://docs.google.com/document/d/1kvl_hwnIG3lAXcSxRndgcqbDGEvGoiD52-VwOuVfveE/edit?resourcekey=0-uvlZUrkCQ-o0H8ohYepsqA
> > To do so we are changing viz, cc, and other chrome parts to use
> > SkColor4f instead of SkColor.
> >
> > PaintFlags already stores SkColor4f instead of SkColor, this CL is only
> > changing the getAlpha and setAlpha interface, as to make explicit from
> > users of the code that alpha is a float now. This CL saves some
> > conversions between SkColor4f and SkColor that shouldn't be necessary.
> >
> > The next steps for PaintFlags would be to also remove the setColor that
> > expects SkColor.
> > The next steps for some of the callers of PaintFlags::getAlpha/setAlpha
> > would be to make them use the float received without having to convert
> > it to int.
> >
> > Bug: 1308932
> > Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3696922
> > Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
> > Cr-Commit-Position: refs/heads/main@{#1016997}
>
> Bug: 1308932
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3721246
> Commit-Queue: Juanmi Huertas <juanmihd@chromium.org>
> Cr-Commit-Position: refs/heads/main@{#1022247}

Bug: 1308932
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3753324
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: Scott Little <sclittle@chromium.org>
Owners-Override: Scott Little <sclittle@chromium.org>
Commit-Queue: Scott Little <sclittle@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1022322}

+++ ui/views/window/frame_caption_button.h
@@ -110,7 +110,7 @@ class VIEWS_EXPORT FrameCaptionButton : public views::Button {
 
   // Determines what alpha to use for the icon based on animation and
   // active state.
-  float GetAlphaForIcon(float base_alpha) const;
+  int GetAlphaForIcon(int base_alpha) const;
 
   void UpdateInkDropBaseColor();
